Refactor octobot package
