+++ octobot/factories/evaluator_factory.pxd
