+++ octobot/factories/interface_factory.py
@@ -14,12 +14,12 @@
 #  You should have received a copy of the GNU Lesser General Public
 #  License along with this library.
 
-from config import PROJECT_NAME, LONG_VERSION
+from octobot.constants import PROJECT_NAME, LONG_VERSION
 from octobot_interfaces.api.interfaces import create_interface_factory, initialize_global_project_data, is_enabled, \
     start_interfaces, is_enabled_in_backtesting
 from octobot_interfaces.util.bot import get_bot_api
 from octobot_notifications.api.notification import create_notifier_factory, is_enabled_in_config
-from tools import get_logger
+from octobot.logger import get_logger
 
 
 class InterfaceFactory:
