+++ start.py
@@ -18,10 +18,11 @@ import asyncio
 import aiohttp
 import os
 import sys
-import webbrowser
-import socket
-from time import sleep
 
+from octobot.disclaimer import DISCLAIMER
+from octobot.constants import LONG_VERSION, FORCE_ASYNCIO_DEBUG_OPTION
+from octobot.configuration_manager import config_health_check
+from octobot.logger import init_logger
 from octobot_commons.config import load_config, is_config_empty_or_missing, init_config
 from octobot_commons.config_manager import validate_config_file, accepted_terms, is_in_dev_mode
 from octobot_commons.constants import CONFIG_ENABLED_OPTION, CONFIG_FILE, DEFAULT_CONFIG_FILE
@@ -32,19 +33,12 @@ from octobot_tentacles_manager.api.installer import install_all_tentacles
 from octobot_tentacles_manager.cli import register_tentacles_manager_arguments
 from octobot_tentacles_manager.constants import DEFAULT_TENTACLES_URL
 from octobot_trading.constants import CONFIG_TRADER, CONFIG_SIMULATOR, CONFIG_TRADING, CONFIG_TRADER_RISK
-from config import LONG_VERSION, FORCE_ASYNCIO_DEBUG_OPTION
-from tools.commands import exchange_keys_encrypter, start_strategy_optimizer, start_bot, data_collector, \
+from octobot.commands import exchange_keys_encrypter, start_strategy_optimizer, start_bot, data_collector, \
     call_tentacles_manager
-from tools.config_manager import config_health_check
-from tools.logger import init_logger
-
-from config.disclaimer import DISCLAIMER
 from octobot_tentacles_manager.api.loader import load_tentacles
 
 
 # Keep string '+' operator to ensure backward compatibility in this file
-
-
 def update_config_with_args(starting_args, config, logger):
     if starting_args.backtesting:
         try:
